requirements.txt
Flask
Flask-WTF
Flask-Login
Flask-Bcrypt
Flask-SQLAlchemy
email-validator
python-dotenv
